Add CSV export option for niViz
Some typical 300,000 measurement points of an SMP run are often too much
for niViz to handle, so this new call allows to thin out the dataset.
It can reproject to a slope angle and stretch the profile.
It also handles some technical details (distance must be in cm,
no header must be present for niViz).