Automate GUI actions for all parameterizations
Now (hopefully) it is enough to copy and modify the core density and SSA
functions to introduce new parameterizations.
Everything from creating the menus to choosing different plot colors to
exporting all parameterizations (albeit still with the same resolution
and windowing) is now handled on the fly, in effect getting rid of all
GUI overhead.