Prepare export of only one parameterization
To tackle the problem of how to merge parameterizations with different
resolutions (window_size, overlap_factor) the internales are changed a
bit:
- For plotting in the GUI, the derivatives now each have their own
distances with which they calculate (one dataframe per param instead of
a single huge one).
- For exporting, we calculate a single parameterization only. This will
be settable in the preferences.
(This is an intermediate commit -- more work to come)