Add possibility to remove air gap in plots
Next to the superposition-view is now a new icon which will remove
all samples before the "ground" marker (if such a marker exists), as
well as shift everything else (derivatives, markers). If a marker lands
in the negative it will not be displayed (else the axis again will
not have the same x start). The drift is also capped for the same
reason.
Also a small optimization in recalc_derivatives to use a function we
already have.