Add CAAML export window
None of the settings do anything yet, but a CAAML is written out.
For this the niViz export classes have been renamed to make place.
This draft of the new main export window shows a tab for CAAML export
(which will pack every info we have for e. g. niViz or SNOWPACK) and a
tab for the grain type classification we need to implement to be able to
drive snow models.