Put AI building blocks together
Providing a training data folder in the export window we get a CAAML
with classified grains.
For this some business logic was moved to the Document class which will
become more important as an abstraction layer. (In the future this could
cache the derivatives and grain shapes for plots, exports, etc.)