Implement preprocessing algorithms
A simplistic approach is now offered for each of the preprocessing
options (albeit largely untested so far).
Also, instaed of 100s of thousands of samples we output the SMP force
profile in the same resolution as the derivatives.
Not all is set yet but now we get a CAAML that is readable by niViz!