Add pre-processing to training data, refac and doc
The low level pre-processing (like removing negative values) is now
applied to the training data set also.
Add documentation and get rid of passing around the parameterization
when we can easily find it from the dataset.
Some dev code was made into a proper example for the API (i. e. not GUI).