Update views when setting surface marker
When the surface marker is set through the toolbar the superposition
view is now updated. Also, if "hide air gap" is enabled it is
immediately cut off from the view. Note that this currently does not
work for other means of setting (or deleting) the surface marker since
old code must be untangled here beforehand.